Added game
My team at work used to play jeopardy on the Google assistant app after stand up every day. the app was horrible and kept crashing so used your API to build a better version in browser, here it is. Thank you so much for working on this it's really great, when it went down for a few days we had to go back to the app and it was a real pain so really thankful you have this up!!